Add markdown printing, daily comment states
This adds two main features.

We use instructions provided at https://stackoverflow.com/a/32035217 to
provide markdown printing of Python output. This helps to make our
outputs that aren't graphs look pretty and readable.

This also adds some analysis that is based on the daily resampling of
the comment counts, since as comments per day and times between comments